Prevent multiple apps grid collapse animations
This could happen if both animate-app-display and fast-app-close are
enabled, and the user clicks again while the grid close aniamation is
still happening.

Also, fast-app-close made necessary to always hide apps page before
toggling overview, even if animate-app-display is disabled (otherwise
the overview could be hidden before the animation ended).

We now define this.closeAppsPage on extension init, as a unified way
to close the overview from the apps page (regardless of the animation).